+++ ot/sliced/__init__.py
@@ -13,7 +13,9 @@ Solvers related to (balanced) sliced transport.
 from ._utils import (
     get_random_projections,
     get_projections_sphere,
+    get_random_rotations,
     projection_sphere_to_circle,
+    projection_sphere_to_ball,
 )
 from ._sliced_distances import (
     sliced_wasserstein_distance,
@@ -23,13 +25,16 @@ from ._spherical_sliced import (
     sliced_wasserstein_sphere,
     sliced_wasserstein_sphere_unif,
     linear_sliced_wasserstein_sphere,
+    stereographic_sliced_wasserstein_sphere,
 )
 from ._sliced_plans import min_sliced_transport_plan, expected_sliced_plan, sliced_plans
 
 __all__ = [
     "get_random_projections",
     "get_projections_sphere",
+    "get_random_rotations",
     "projection_sphere_to_circle",
+    "projection_sphere_to_ball",
     "min_sliced_transport_plan",
     "expected_sliced_plan",
     "sliced_plans",
@@ -38,4 +43,5 @@ __all__ = [
     "sliced_wasserstein_sphere",
     "sliced_wasserstein_sphere_unif",
     "linear_sliced_wasserstein_sphere",
+    "stereographic_sliced_wasserstein_sphere",
 ]